requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
pandas==2.1.3
matplotlib==3.8.2
python-dotenv==1.0.0
jinja2==3.1.2